(#1) fixed migration

+++ db/migrate/20160904195237_create_time_entries.rb
@@ -3,7 +3,7 @@ class CreateTimeEntries < ActiveRecord::Migration
     create_table :time_entries do |t|
       t.integer :day_id
       t.integer :type_id
-      t.integer :second
+      t.float :second
 
       t.timestamps null: false
     end
